+++ build.gradle
@@ -87,7 +87,6 @@ clean {
     delete file('src/main/generated')
 }
 
-
 jacocoTestCoverageVerification {
     violationRules {
         rule {
