+++ src/main/java/world/trecord/config/redis/RedisConfig.java
@@ -12,7 +12,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
-import world.trecord.config.security.UserContext;
+import world.trecord.config.security.account.UserContext;
 
 @Configuration
 @EnableRedisRepositories
