+++ src/main/java/world/trecord/config/redis/UserCacheRepository.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Repository;
-import world.trecord.config.security.UserContext;
+import world.trecord.config.security.account.UserContext;
 
 import java.time.Duration;
 import java.util.Optional;
