+++ src/main/java/world/trecord/config/security/account/AccountContext.java
@@ -1,4 +1,4 @@
-package world.trecord.config.security;
+package world.trecord.config.security.account;
 
 public interface AccountContext {
     Long getId();
