+++ src/main/java/world/trecord/config/security/account/AccountRole.java
@@ -1,0 +1,5 @@
+package world.trecord.config.security.account;
+
+public enum AccountRole {
+    ROLE_USER;
+}
