+++ src/main/java/world/trecord/config/security/account/CurrentContext.java
@@ -1,4 +1,4 @@
-package world.trecord.config.security;
+package world.trecord.config.security.account;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
