+++ src/main/java/world/trecord/controller/notification/NotificationController.java
@@ -8,9 +8,9 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-import world.trecord.config.security.AccountContext;
-import world.trecord.config.security.CurrentContext;
-import world.trecord.config.security.UserContext;
+import world.trecord.config.security.account.AccountContext;
+import world.trecord.config.security.account.CurrentContext;
+import world.trecord.config.security.account.UserContext;
 import world.trecord.controller.ApiResponse;
 import world.trecord.domain.notification.enumeration.NotificationType;
 import world.trecord.dto.notification.response.CheckNewNotificationResponse;
