+++ src/main/java/world/trecord/controller/users/UserController.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
-import world.trecord.config.security.AccountContext;
-import world.trecord.config.security.CurrentContext;
+import world.trecord.config.security.account.AccountContext;
+import world.trecord.config.security.account.CurrentContext;
 import world.trecord.controller.ApiResponse;
 import world.trecord.dto.comment.response.UserCommentResponse;
 import world.trecord.dto.feedcontributor.response.UserFeedContributorListResponse;
