--- src/main/java/world/trecord/domain/users/Role.java
@@ -1,5 +1,0 @@
-package world.trecord.domain.users;
-
-public enum Role {
-    ROLE_USER;
-}
